Cambio de costo a tipo flotante

+++ database/migrations/2020_11_10_225409_piezas_refaccion.php
@@ -18,7 +18,7 @@ class PiezasRefaccion extends Migration
             $table->string('nombre');
             $table->string('descripcion');
             $table->integer('piezasDisponibles');
-            $table->integer('costo');
+            $table->float('costo', 8, 2);
             $table->timestamps();
         });
     }
